+++ tests/fixtures/parser/ast/svelte5/attach-ts-01-output.json
@@ -1939,124 +1939,16 @@
       }
     },
     {
-      "type": "Punctuator",
-      "value": "@",
+      "type": "MustacheKeyword",
+      "value": "@attach",
       "range": [
         242,
-        243
-      ],
-      "loc": {
-        "start": {
-          "line": 14,
-          "column": 6
-        },
-        "end": {
-          "line": 14,
-          "column": 7
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        243,
-        244
-      ],
-      "loc": {
-        "start": {
-          "line": 14,
-          "column": 7
-        },
-        "end": {
-          "line": 14,
-          "column": 8
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        244,
-        245
-      ],
-      "loc": {
-        "start": {
-          "line": 14,
-          "column": 8
-        },
-        "end": {
-          "line": 14,
-          "column": 9
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        245,
-        246
-      ],
-      "loc": {
-        "start": {
-          "line": 14,
-          "column": 9
-        },
-        "end": {
-          "line": 14,
-          "column": 10
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        246,
-        247
-      ],
-      "loc": {
-        "start": {
-          "line": 14,
-          "column": 10
-        },
-        "end": {
-          "line": 14,
-          "column": 11
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "c",
-      "range": [
-        247,
-        248
-      ],
-      "loc": {
-        "start": {
-          "line": 14,
-          "column": 11
-        },
-        "end": {
-          "line": 14,
-          "column": 12
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "h",
-      "range": [
-        248,
         249
       ],
       "loc": {
         "start": {
           "line": 14,
-          "column": 12
+          "column": 6
         },
         "end": {
           "line": 14,
