+++ tests/fixtures/parser/ast/svelte5/attach-ts-02-output.json
@@ -2777,124 +2777,16 @@
       }
     },
     {
-      "type": "Punctuator",
-      "value": "@",
+      "type": "MustacheKeyword",
+      "value": "@attach",
       "range": [
         348,
-        349
-      ],
-      "loc": {
-        "start": {
-          "line": 17,
-          "column": 9
-        },
-        "end": {
-          "line": 17,
-          "column": 10
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        349,
-        350
-      ],
-      "loc": {
-        "start": {
-          "line": 17,
-          "column": 10
-        },
-        "end": {
-          "line": 17,
-          "column": 11
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        350,
-        351
-      ],
-      "loc": {
-        "start": {
-          "line": 17,
-          "column": 11
-        },
-        "end": {
-          "line": 17,
-          "column": 12
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        351,
-        352
-      ],
-      "loc": {
-        "start": {
-          "line": 17,
-          "column": 12
-        },
-        "end": {
-          "line": 17,
-          "column": 13
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        352,
-        353
-      ],
-      "loc": {
-        "start": {
-          "line": 17,
-          "column": 13
-        },
-        "end": {
-          "line": 17,
-          "column": 14
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "c",
-      "range": [
-        353,
-        354
-      ],
-      "loc": {
-        "start": {
-          "line": 17,
-          "column": 14
-        },
-        "end": {
-          "line": 17,
-          "column": 15
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "h",
-      "range": [
-        354,
         355
       ],
       "loc": {
         "start": {
           "line": 17,
-          "column": 15
+          "column": 9
         },
         "end": {
           "line": 17,
