+++ tests/fixtures/parser/ast/svelte5/attach-ts-03-output.json
@@ -1888,124 +1888,16 @@
       }
     },
     {
-      "type": "Punctuator",
-      "value": "@",
+      "type": "MustacheKeyword",
+      "value": "@attach",
       "range": [
         105,
-        106
-      ],
-      "loc": {
-        "start": {
-          "line": 8,
-          "column": 2
-        },
-        "end": {
-          "line": 8,
-          "column": 3
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        106,
-        107
-      ],
-      "loc": {
-        "start": {
-          "line": 8,
-          "column": 3
-        },
-        "end": {
-          "line": 8,
-          "column": 4
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        107,
-        108
-      ],
-      "loc": {
-        "start": {
-          "line": 8,
-          "column": 4
-        },
-        "end": {
-          "line": 8,
-          "column": 5
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        108,
-        109
-      ],
-      "loc": {
-        "start": {
-          "line": 8,
-          "column": 5
-        },
-        "end": {
-          "line": 8,
-          "column": 6
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        109,
-        110
-      ],
-      "loc": {
-        "start": {
-          "line": 8,
-          "column": 6
-        },
-        "end": {
-          "line": 8,
-          "column": 7
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "c",
-      "range": [
-        110,
-        111
-      ],
-      "loc": {
-        "start": {
-          "line": 8,
-          "column": 7
-        },
-        "end": {
-          "line": 8,
-          "column": 8
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "h",
-      "range": [
-        111,
         112
       ],
       "loc": {
         "start": {
           "line": 8,
-          "column": 8
+          "column": 2
         },
         "end": {
           "line": 8,
