+++ tests/fixtures/parser/ast/svelte5/attach01-output.json
@@ -1393,124 +1393,16 @@
       }
     },
     {
-      "type": "Punctuator",
-      "value": "@",
+      "type": "MustacheKeyword",
+      "value": "@attach",
       "range": [
         215,
-        216
-      ],
-      "loc": {
-        "start": {
-          "line": 12,
-          "column": 6
-        },
-        "end": {
-          "line": 12,
-          "column": 7
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        216,
-        217
-      ],
-      "loc": {
-        "start": {
-          "line": 12,
-          "column": 7
-        },
-        "end": {
-          "line": 12,
-          "column": 8
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        217,
-        218
-      ],
-      "loc": {
-        "start": {
-          "line": 12,
-          "column": 8
-        },
-        "end": {
-          "line": 12,
-          "column": 9
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        218,
-        219
-      ],
-      "loc": {
-        "start": {
-          "line": 12,
-          "column": 9
-        },
-        "end": {
-          "line": 12,
-          "column": 10
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        219,
-        220
-      ],
-      "loc": {
-        "start": {
-          "line": 12,
-          "column": 10
-        },
-        "end": {
-          "line": 12,
-          "column": 11
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "c",
-      "range": [
-        220,
-        221
-      ],
-      "loc": {
-        "start": {
-          "line": 12,
-          "column": 11
-        },
-        "end": {
-          "line": 12,
-          "column": 12
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "h",
-      "range": [
-        221,
         222
       ],
       "loc": {
         "start": {
           "line": 12,
-          "column": 12
+          "column": 6
         },
         "end": {
           "line": 12,
