+++ tests/fixtures/parser/ast/svelte5/attach02-output.json
@@ -2253,124 +2253,16 @@
       }
     },
     {
-      "type": "Punctuator",
-      "value": "@",
+      "type": "MustacheKeyword",
+      "value": "@attach",
       "range": [
         356,
-        357
-      ],
-      "loc": {
-        "start": {
-          "line": 20,
-          "column": 9
-        },
-        "end": {
-          "line": 20,
-          "column": 10
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        357,
-        358
-      ],
-      "loc": {
-        "start": {
-          "line": 20,
-          "column": 10
-        },
-        "end": {
-          "line": 20,
-          "column": 11
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        358,
-        359
-      ],
-      "loc": {
-        "start": {
-          "line": 20,
-          "column": 11
-        },
-        "end": {
-          "line": 20,
-          "column": 12
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "t",
-      "range": [
-        359,
-        360
-      ],
-      "loc": {
-        "start": {
-          "line": 20,
-          "column": 12
-        },
-        "end": {
-          "line": 20,
-          "column": 13
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "a",
-      "range": [
-        360,
-        361
-      ],
-      "loc": {
-        "start": {
-          "line": 20,
-          "column": 13
-        },
-        "end": {
-          "line": 20,
-          "column": 14
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "c",
-      "range": [
-        361,
-        362
-      ],
-      "loc": {
-        "start": {
-          "line": 20,
-          "column": 14
-        },
-        "end": {
-          "line": 20,
-          "column": 15
-        }
-      }
-    },
-    {
-      "type": "Identifier",
-      "value": "h",
-      "range": [
-        362,
         363
       ],
       "loc": {
         "start": {
           "line": 20,
-          "column": 15
+          "column": 9
         },
         "end": {
           "line": 20,
