+++ tests/fixtures/parser/ast/svelte5/attach04-input.svelte
@@ -1,9 +1,23 @@
 <script>
-	/** @type {import('svelte/elements').HTMLButtonAttributes} */
-	let { children, ...props } = $props();
+	import tippy from 'tippy.js';
+	import Button from './Button.svelte';
+
+	let content = $state('Hello!');
+
+	/**
+	 * @param {string} content
+	 * @returns {import('svelte/attachments').Attachment}
+	 */
+	function tooltip(content) {
+		return (element) => {
+			const tooltip = tippy(element, { content });
+			return tooltip.destroy;
+		};
+	}
 </script>
 
-<!-- `props` includes attachments -->
-<button {...props}>
-	{@render children?.()}
-</button>
+<input bind:value={content} />
+
+<Button {@attach tooltip(content)}>
+	Hover me
+</Button>
