--- tests/fixtures/parser/ast/svelte5/attach04-prefer-const-result.json
@@ -1,14 +1,0 @@
-[
-  {
-    "ruleId": "prefer-const",
-    "code": "children",
-    "line": 3,
-    "column": 8
-  },
-  {
-    "ruleId": "prefer-const",
-    "code": "props",
-    "line": 3,
-    "column": 21
-  }
-]
