--- tests/fixtures/parser/ast/svelte5/attach04-ts-prefer-const-result.json
@@ -1,14 +1,0 @@
-[
-  {
-    "ruleId": "prefer-const",
-    "code": "children",
-    "line": 4,
-    "column": 8
-  },
-  {
-    "ruleId": "prefer-const",
-    "code": "props",
-    "line": 4,
-    "column": 21
-  }
-]
