--- tests/fixtures/parser/ast/svelte5/attach05-input.svelte
@@ -1,23 +1,0 @@
-<script>
-	import tippy from 'tippy.js';
-	import Button from './Button.svelte';
-
-	let content = $state('Hello!');
-
-	/**
-	 * @param {string} content
-	 * @returns {import('svelte/attachments').Attachment}
-	 */
-	function tooltip(content) {
-		return (element) => {
-			const tooltip = tippy(element, { content });
-			return tooltip.destroy;
-		};
-	}
-</script>
-
-<input bind:value={content} />
-
-<Button {@attach tooltip(content)}>
-	Hover me
-</Button>
